requirements.txt
django==4.0.5
python-dateutil==2.8.2